+++ bootstrap/app.php
@@ -5,7 +5,7 @@ use Illuminate\Foundation\Configuration\Exceptions;
 use Illuminate\Foundation\Configuration\Middleware;
 // use Illuminate\Http\Request;
 
-return Application::configure(basePath: dirname(__DIR__))
+$app = Application::configure(basePath: dirname(__DIR__))
     ->withRouting(
         web: __DIR__.'/../routes/web.php',
         commands: __DIR__.'/../routes/console.php',
@@ -34,3 +34,14 @@ return Application::configure(basePath: dirname(__DIR__))
     ->withExceptions(function (Exceptions $exceptions): void {
         // ?
     })->create();
+
+// === CARREGAMENTO PRECOCE DO .ENV ===
+// Como o autoloader já rodou, podemos usar o Dotenv nativo do Laravel
+if (file_exists(__DIR__.'/../.env')) {
+    \Dotenv\Dotenv::createUnsafeImmutable(__DIR__.'/..')->safeLoad();
+}
+
+$publicPath = env('APP_PUBLIC_PATH') ?: $app->publicPath();
+$app->usePublicPath($publicPath);
+
+return $app;